requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
pydantic==2.5.0
asyncio-mqtt==0.16.1
authlib==1.2.1
tensorflow==2.15.0
scikit-learn==1.3.2
joblib==1.3.2
